[tracegen] move component under cmd
Fixes #12474
