+++ cmd/tracegen/internal/tracegen/config_test.go
